requirements.txt
alembic==1.8.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
Flask-Admin==1.6.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
PyMySQL==1.0.2
python-slugify==6.1.2
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.37
text-unidecode==1.3
uni-slugify==0.1.4
Unidecode==1.3.4
urllib3==1.26.9
WebHelpers==1.3
Werkzeug==2.1.2
WTForms==3.0.1
xmltodict==0.13.0
zipp==3.8.0
